feat(backdrop): add backdrop support
A component has been added to the default layout, which displays an image from a Vuex store. The
Vuex store can either set or clear backdrops. To set a backdrop, you must pass a BaseItemDto[] in
order to retrieve the relevant images. It is designed to allow a rotation of the array, in order to
change which backdrop is displayed.